simplify the [game] routing, scaffold out [game]/[mod] route
