+++ src/routes/[game]/+layout.svelte
