+++ src/routes/[game]/[mod]/+layout.svelte
@@ -1,0 +1,7 @@
+<script lang="ts">
+  import type { LayoutProps } from "./$types";
+
+  let { data, children }: LayoutProps = $props();
+</script>
+
+{@render children()}
