+++ src/routes/[game]/[mod]/+page.svelte
@@ -1,5 +1,7 @@
 <script lang="ts">
-  import type { PageProps } from "./$types";
+  import type { PageProps } from "./mod/$types";
 
   let { data }: PageProps = $props();
 </script>
+
+mod page
