+++ src/routes/[game]/[mod]/GameControlsMod.svelte
