+++ src/routes/[game]/textures/+page.svelte
